Put CameraAnimatorsWeakSet in separate file
